config: move CLI helpers into config
